+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.1-alpha] - 2019-09-02
+### Added
+- Exchange global access through Exchanges singleton
+
+### Fixed
+- Trader enabled method call in __init__
+
 ## [1.3.0-alpha] - 2019-09-01
 ### Added
 - Trading mode implementation